+++ TrabGB/DiamondView.h
@@ -3,7 +3,7 @@
 #include "Player.h"
 class DiamondView {
 public:
-	DiamondView(int defaultTileHeight, int defaultTileWidth, int numberOfTilesVertically, int numberOfTilesHorizontally);
+	DiamondView(float defaultTileHeight, float defaultTileWidth, int numberOfTilesVertically, int numberOfTilesHorizontally);
 	glm::vec2 tileWalking(int r, int c, int direction);
 	glm::vec2 tileWalkingToCorrectDirection(int r, int c);
 	glm::vec2 getRowAndColumnForMousePositionClick(int x, int y);
